@@ -8,11 +8,15 @@
     "@mui/icons-material": "^5.8.0",
     "@mui/material": "^5.8.0",
     "@mui/styles": "^5.7.0",
+    "axios": "^1.7.2",
+    "bootstrap": "^5.3.3",
     "moment": "^2.29.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-moment": "^1.1.2",
+    "react-router-dom": "^6.23.1",
     "react-scripts": "5.0.0",
+    "react-validation": "^3.0.7",
     "typescript": "^4.6.4"
   },
   "scripts": {
