@@ -15,6 +15,6 @@
  */
 // Copy from the endpoint from the API Gateway Deployment
 // Example: const API_LIST = 'https://di2eyonlz5s7kmuektcddaw5zq.apigateway.<region>.oci.customer-oci.com/todolist';
-const API_LIST = 'http://localhost:8085/api/tarea';
+const API_LIST = 'http://localhost:8085/api';
 //const API_LIST = '/todolist';
 export default API_LIST;
